Simplifie la copie dans le presse-papier
